refactor(tools): separate enhancement inspection from rewriting
